tests work with postgres and env variables
